+++ src/main/kotlin/me/blvckbytes/aoc_2023/data/ScratchCard.kt
@@ -1,7 +1,7 @@
 package me.blvckbytes.aoc_2023.data
 
 class ScratchCard(
-  val winningNumbers: Set<Int>,
-  val ownNumbers: Set<Int>,
-  val matchingNumbers: Set<Int>
+  val winningNumbers: Set<Long>,
+  val ownNumbers: Set<Long>,
+  val matchingNumbers: Set<Long>
 )
